move rle to own file
