+++ crates/sourisdb/src/types/binary.rs
@@ -4,16 +4,18 @@ use crate::{
     utilities::cursor::Cursor,
     values::ValueTy,
 };
-use alloc::{vec, vec::Vec};
+use alloc::{vec::Vec};
 use core::fmt::{Debug, Display, Formatter};
-use itertools::Itertools;
 use serde_json::{Map as SJMap, Number, Value as SJValue};
+use crate::types::binary::rle::{rle, un_rle};
+
+pub mod rle;
 
 #[derive(Debug, Copy, Clone)]
 pub enum BinaryCompression {
     Nothing,
     RunLengthEncoding,
-    XORedRunLengthEncoding,
+    LempelZiv, //basic algo taken only from: https://go-compression.github.io/algorithms/lz/. I did not go to the next page
 }
 
 impl From<BinaryCompression> for u8 {
@@ -21,7 +23,7 @@ impl From<BinaryCompression> for u8 {
         match compression {
             BinaryCompression::Nothing => 0,
             BinaryCompression::RunLengthEncoding => 1,
-            BinaryCompression::XORedRunLengthEncoding => 2,
+            BinaryCompression::LempelZiv => 2,
         }
     }
 }
@@ -33,7 +35,7 @@ impl TryFrom<u8> for BinaryCompression {
         match value {
             0 => Ok(Self::Nothing),
             1 => Ok(Self::RunLengthEncoding),
-            2 => Ok(Self::XORedRunLengthEncoding),
+            2 => Ok(Self::LempelZiv),
             _ => Err(BinarySerError::NoCompressionTypeFound(value)),
         }
     }
@@ -113,76 +115,6 @@ impl BinaryData {
         SJValue::Object(obj)
     }
 
-    #[must_use]
-    pub fn rle(&self) -> Vec<u8> {
-        let mut output = vec![];
-
-        if self.0.is_empty() {
-            return output;
-        }
-
-        let mut bytes = self.0.clone();
-        bytes.reverse();
-
-        let mut current_count = 1;
-        let mut current = bytes.pop().unwrap();
-
-        while let Some(byte) = bytes.pop() {
-            if current != byte {
-                output.push(current_count);
-                output.push(current);
-
-                current_count = 0;
-                current = byte;
-            }
-            current_count += 1;
-
-            if current_count == u8::MAX {
-                output.push(current_count);
-                output.push(current);
-
-                current_count = 0;
-            }
-        }
-        if current_count != 0 {
-            output.push(current_count);
-            output.push(current);
-        }
-
-        output
-    }
-
-    fn un_rle(len: usize, cursor: &mut Cursor<u8>) -> Result<Self, BinarySerError> {
-        //complicated version that's like 70% slower lol
-        /*        let mut alloc_size = 0;
-                let mut to_be_added = Vec::with_capacity(len);
-
-                for _ in 0..len {
-                    let [count, byte] = cursor.read_exact().copied().ok_or(BinarySerError::NotEnoughBytes)?;
-                    alloc_size += count as usize;
-                    to_be_added.push((count, byte));
-                }
-
-                let mut output = Vec::with_capacity(alloc_size);
-
-                for (count, byte) in to_be_added {
-                    (0..count).for_each(|_| output.push(byte));
-                }
-        */
-
-        let mut output = vec![];
-        for (count, byte) in cursor
-            .read(len * 2)
-            .ok_or(BinarySerError::NotEnoughBytes)?
-            .iter()
-            .copied()
-            .tuples()
-        {
-            (0..count).for_each(|_| output.push(byte));
-        }
-
-        Ok(Self(output))
-    }
 
     #[must_use]
     pub fn ser(&self) -> (BinaryCompression, Vec<u8>) {
@@ -192,15 +124,12 @@ impl BinaryData {
             backing
         };
         let rle = {
-            let rle = self.rle();
+            let rle = rle(self.0.clone());
 
             let mut out = Integer::usize(rle.len() / 2).ser().1;
             out.extend(&rle);
             out
         };
-        // let rle_xor = {
-        //TODO eventually
-        // };
 
         if vanilla.len() <= rle.len() {
             (BinaryCompression::Nothing, vanilla)
@@ -222,45 +151,10 @@ impl BinaryData {
                     .ok_or(BinarySerError::NotEnoughBytes)?
                     .to_vec(),
             ),
-            BinaryCompression::RunLengthEncoding => Self::un_rle(len, cursor)?,
-            BinaryCompression::XORedRunLengthEncoding => {
+            BinaryCompression::RunLengthEncoding => Self(un_rle(len, cursor)?),
+            BinaryCompression::LempelZiv => {
                 unimplemented!()
             }
         })
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_rle_specific_cases() {
-        const CASES: &[&[u8]] = &[
-            &[
-                0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0xFF,
-            ],
-            &[],
-            &[0],
-            &[0x12, 0x12, 0x12, 0xDE, 0xAD, 0xBE, 0xEF],
-        ];
-
-        for case in CASES {
-            let vec = case.to_vec();
-            let bd = BinaryData(vec.clone());
-
-            let encoded = {
-                let rle = bd.rle(); //forcing RLE to ensure it works
-                let mut out = Integer::usize(rle.len() / 2).ser().1;
-                out.extend(&rle);
-                out
-            };
-
-            let mut cursor = Cursor::new(&encoded);
-            let BinaryData(decoded) =
-                BinaryData::deser(BinaryCompression::RunLengthEncoding, &mut cursor).unwrap();
-
-            assert_eq!(decoded, vec);
-        }
     }
 }
